fix postgres table creation

+++ app/database/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
-from sqlalchemy import String, Integer, DateTime, Text, ForeignKey
+from sqlalchemy import String, Integer, BigInteger, DateTime, Text, ForeignKey
 
 class Base(DeclarativeBase):
     pass
@@ -8,7 +8,7 @@ class Base(DeclarativeBase):
 class Income(Base):
     __tablename__ = 'incomes'
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    amount: Mapped[int] = mapped_column(Integer)
+    amount: Mapped[int] = mapped_column(BigInteger)
     source: Mapped[str] = mapped_column(String(120))
     note: Mapped[str] = mapped_column(Text, default='')
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
@@ -16,7 +16,7 @@ class Income(Base):
 class Expense(Base):
     __tablename__ = 'expenses'
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    amount: Mapped[int] = mapped_column(Integer)
+    amount: Mapped[int] = mapped_column(BigInteger)
     category: Mapped[str] = mapped_column(String(120))
     note: Mapped[str] = mapped_column(Text, default='')
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
@@ -29,7 +29,7 @@ class Debt(Base):
     # person: kimdan qarz oldim yoki kimga qarz berdim
     person: Mapped[str] = mapped_column(String(150), default='')
     name: Mapped[str] = mapped_column(String(150))
-    total_amount: Mapped[int] = mapped_column(Integer)
+    total_amount: Mapped[int] = mapped_column(BigInteger)
     note: Mapped[str] = mapped_column(Text, default='')
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
 
@@ -37,6 +37,6 @@ class DebtPayment(Base):
     __tablename__ = 'debt_payments'
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
     debt_id: Mapped[int] = mapped_column(ForeignKey('debts.id'))
-    amount: Mapped[int] = mapped_column(Integer)
+    amount: Mapped[int] = mapped_column(BigInteger)
     note: Mapped[str] = mapped_column(Text, default='')
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
